+++ HISTORY.md
@@ -1,4 +1,10 @@
 
+## Next release
+
+### Changed
+
+- Interface of oph_gsl_spline
+
 ## v1.8.0 - 2023-07-31
 
 ### Fixed